backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
python-dotenv==1.0.0
python-dateutil==2.8.2
openai>=1.12.0
slowapi>=0.1.9
